Move parts of the triggering logic into the widget
