+++ src/vs/editor/contrib/hover/hover.ts
@@ -7,7 +7,6 @@ import * as nls from 'vs/nls';
 import { IKeyboardEvent } from 'vs/base/browser/keyboardEvent';
 import { KeyChord, KeyCode, KeyMod } from 'vs/base/common/keyCodes';
 import { IDisposable, DisposableStore } from 'vs/base/common/lifecycle';
-import { IEmptyContentData } from 'vs/editor/browser/controller/mouseTarget';
 import { ICodeEditor, IEditorMouseEvent, MouseTargetType } from 'vs/editor/browser/editorBrowser';
 import { EditorAction, ServicesAccessor, registerEditorAction, registerEditorContribution } from 'vs/editor/browser/editorExtensions';
 import { ConfigurationChangedEvent, EditorOption } from 'vs/editor/common/config/editorOptions';
@@ -165,45 +164,32 @@ export class ModesHoverController implements IEditorContribution {
 			return;
 		}
 
-		if (targetType === MouseTargetType.CONTENT_EMPTY) {
-			const epsilon = this._editor.getOption(EditorOption.fontInfo).typicalHalfwidthCharacterWidth / 2;
-			const data = <IEmptyContentData>mouseEvent.target.detail;
-			if (data && !data.isAfterLines && typeof data.horizontalDistanceToText === 'number' && data.horizontalDistanceToText < epsilon) {
-				// Let hover kick in even when the mouse is technically in the empty area after a line, given the distance is small enough
-				targetType = MouseTargetType.CONTENT_TEXT;
-			}
+		if (!this._isHoverEnabled) {
+			this._hideWidgets();
+			return;
 		}
 
-		if (targetType === MouseTargetType.CONTENT_TEXT) {
+		const contentWidget = this._getOrCreateContentWidget();
+		const contentHoverTrigger = contentWidget.checkTrigger(mouseEvent);
+		if (contentHoverTrigger) {
 			this._glyphWidget?.hide();
+			contentWidget.startShowingAt(contentHoverTrigger.range, HoverStartMode.Delayed, false);
+			return;
+		}
 
-			if (this._isHoverEnabled && mouseEvent.target.range) {
-				// TODO@rebornix. This should be removed if we move Color Picker out of Hover component.
-				// Check if mouse is hovering on color decorator
-				const hoverOnColorDecorator = [...mouseEvent.target.element?.classList.values() || []].find(className => className.startsWith('ced-colorBox'))
-					&& mouseEvent.target.range.endColumn - mouseEvent.target.range.startColumn === 1;
-				const showAtRange = (
-					hoverOnColorDecorator // shift the mouse focus by one as color decorator is a `before` decoration of next character.
-						? new Range(mouseEvent.target.range.startLineNumber, mouseEvent.target.range.startColumn + 1, mouseEvent.target.range.endLineNumber, mouseEvent.target.range.endColumn + 1)
-						: mouseEvent.target.range
-				);
-				if (!this._contentWidget) {
-					this._contentWidget = new ModesContentHoverWidget(this._editor, this._hoverVisibleKey, this._instantiationService);
-				}
-				this._contentWidget.startShowingAt(showAtRange, HoverStartMode.Delayed, false);
-			}
-		} else if (targetType === MouseTargetType.GUTTER_GLYPH_MARGIN) {
+		if (targetType === MouseTargetType.GUTTER_GLYPH_MARGIN) {
 			this._contentWidget?.hide();
 
-			if (this._isHoverEnabled && mouseEvent.target.position) {
+			if (mouseEvent.target.position) {
 				if (!this._glyphWidget) {
 					this._glyphWidget = new ModesGlyphHoverWidget(this._editor, this._modeService, this._openerService);
 				}
 				this._glyphWidget.startShowingAt(mouseEvent.target.position.lineNumber);
+				return;
 			}
-		} else {
-			this._hideWidgets();
 		}
+
+		this._hideWidgets();
 	}
 
 	private _onKeyDown(e: IKeyboardEvent): void {
@@ -223,15 +209,19 @@ export class ModesHoverController implements IEditorContribution {
 		this._contentWidget?.hide();
 	}
 
+	private _getOrCreateContentWidget(): ModesContentHoverWidget {
+		if (!this._contentWidget) {
+			this._contentWidget = new ModesContentHoverWidget(this._editor, this._hoverVisibleKey, this._instantiationService);
+		}
+		return this._contentWidget;
+	}
+
 	public isColorPickerVisible(): boolean {
 		return this._contentWidget?.isColorPickerVisible() || false;
 	}
 
 	public showContentHover(range: Range, mode: HoverStartMode, focus: boolean): void {
-		if (!this._contentWidget) {
-			this._contentWidget = new ModesContentHoverWidget(this._editor, this._hoverVisibleKey, this._instantiationService);
-		}
-		this._contentWidget.startShowingAt(range, mode, focus);
+		this._getOrCreateContentWidget().startShowingAt(range, mode, focus);
 	}
 
 	public dispose(): void {
